requirements.txt
flask
deep-translator
requests
pillow
